[New]Added test cases for pages admin locations list
